Separate the functionality
